:bug: Avoid circular reference between fileUtils and CodecsUtils, which will cause compilation failure in iOS environment
