+++ composeApp/src/commonMain/kotlin/com/crosspaste/utils/StringUtils.kt
@@ -1,0 +1,9 @@
+package com.crosspaste.utils
+
+fun StringBuilder.appendHex(value: ULong) {
+    for (i in 0 until 8) {
+        val byte = (value shr i * 8).toByte()
+        append(HEX_DIGITS[(byte.toInt() shr 4) and 0xf])
+        append(HEX_DIGITS[byte.toInt() and 0xf])
+    }
+}
